# Optimize Scalar Implementation, targeting `aarch64` CPU