Use `__int128` compiler built-in type ( not part of ISO C++ standard ), when available, for performing `u64 x u64 -> u128`

> **Note** This turns out to be pretty fast on ARM64-A profile CPU ! How fast ?

## On ARM Neoverse-V1 ( aka AWS Graviton3 )

At commit 8288a9ec5d04c19f7f0d77686aaef90f9561a3a8 ( **before** )

```bash
2022-12-21T10:06:07+00:00
Running ./bench/a.out
Run on (64 X 2100 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB (x64)
  L1 Instruction 64 KiB (x64)
  L2 Unified 1024 KiB (x64)
  L3 Unified 32768 KiB (x1)
Load Average: 0.18, 0.07, 0.02
-------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations items_per_second max_exec_time (ns) median_exec_time (ns) min_exec_time (ns)
-------------------------------------------------------------------------------------------------------------------------------------------------------------
bench_rphash::permutation/manual_time      26941 ns        41640 ns        25989       37.1182k/s             45.61k               26.897k            26.803k
```

At commit 6c56064b20350b954fa54139d30171a98277efcb ( **after** ) 🎉

```bash
2022-12-21T10:05:04+00:00
Running ./bench/a.out
Run on (64 X 2100 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB (x64)
  L1 Instruction 64 KiB (x64)
  L2 Unified 1024 KiB (x64)
  L3 Unified 32768 KiB (x1)
Load Average: 0.08, 0.03, 0.01
-------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations items_per_second max_exec_time (ns) median_exec_time (ns) min_exec_time (ns)
-------------------------------------------------------------------------------------------------------------------------------------------------------------
bench_rphash::permutation/manual_time      17054 ns        31803 ns        41045       58.6387k/s            28.675k               17.025k            16.946k
```